# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewsDetailView": {
        "properties": {
          "article": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsArticleView"
              },
              {
                "type": "null"
              }
            ]
          },
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "editorial_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Editorial Score"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/NewsMentionView"
            },
            "title": "Mentions",
            "type": "array"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NewsTopicRefView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "source_name",
          "market",
          "fetched_at",
          "mentions"
        ],
        "title": "NewsDetailView",
        "type": "object"
      },
      "NewsEventDetailView": {
        "properties": {
          "event_key": {
            "title": "Event Key",
            "type": "string"
          },
          "event_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Summary"
          },
          "event_title": {
            "title": "Event Title",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "importance_score": {
            "title": "Importance Score",
            "type": "number"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "news_count": {
            "title": "News Count",
            "type": "integer"
          },
          "news_items": {
            "items": {
              "$ref": "#/components/schemas/NewsItemSummary"
            },
            "title": "News Items",
            "type": "array"
          },
          "primary_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Symbol"
          },
          "related_symbols": {
            "items": {
              "type": "string"
            },
            "title": "Related Symbols",
            "type": "array"
          },
          "sentiment_label": {
            "title": "Sentiment Label",
            "type": "string"
          },
          "source_count": {
            "title": "Source Count",
            "type": "integer"
          },
          "watchlist_hits": {
            "items": {
              "type": "string"
            },
            "title": "Watchlist Hits",
            "type": "array"
          }
        },
        "required": [
          "event_key",
          "event_title",
          "event_type",
          "market",
          "sentiment_label",
          "importance_score",
          "related_symbols",
          "source_count",
          "news_count",
          "news_items"
        ],
        "title": "NewsEventDetailView",
        "type": "object"
      },
...
      "NewsFeedLayoutView": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/NewsFeedEventCardView"
            },
            "title": "Events",
            "type": "array"
          },
          "stream": {
            "items": {
              "$ref": "#/components/schemas/NewsItemSummary"
            },
            "title": "Stream",
            "type": "array"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/NewsFeedTopicView"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "events",
          "topics",
          "stream"
        ],
        "title": "NewsFeedLayoutView",
        "type": "object"
      },
...
      "NewsItemSummary": {
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "editorial_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Editorial Score"
          },
          "fetched_at": {
            "title": "Fetched At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "sentiment_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Label"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "source_name",
          "market",
          "fetched_at"
        ],
        "title": "NewsItemSummary",
        "type": "object"
      },
      "NewsListPageView": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NewsItemSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "NewsListPageView",
        "type": "object"
      },
...
      "NewsRefreshResponse": {
        "properties": {
          "fetched_count": {
            "title": "Fetched Count",
            "type": "integer"
          },
          "finished_at": {
            "title": "Finished At",
            "type": "string"
          },
          "inserted_count": {
            "title": "Inserted Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SourceFetchResultView"
            },
            "title": "Results",
            "type": "array"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "started_at",
          "finished_at",
          "fetched_count",
          "inserted_count",
          "results"
        ],
        "title": "NewsRefreshResponse",
        "type": "object"
      },
...
      "NewsRuntimeView": {
        "properties": {
          "degraded_market_count": {
            "title": "Degraded Market Count",
            "type": "integer"
          },
          "feed_status": {
            "title": "Feed Status",
            "type": "string"
          },
          "last_incremental_event_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Incremental Event At"
          },
          "last_news_created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last News Created At"
          },
          "last_refresh_finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Finished At"
          },
          "markets": {
            "items": {
              "$ref": "#/components/schemas/NewsRuntimeMarketView"
            },
            "title": "Markets",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/NewsRuntimeSourceView"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "feed_status",
          "degraded_market_count",
          "markets",
          "sources"
        ],
        "title": "NewsRuntimeView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "News Caught Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/news": {
      "get": {
        "operationId": "list_news_api_news_get",
        "parameters": [
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "source_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Name"
            }
          },
          {
            "in": "query",
            "name": "sentiment_label",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sentiment Label"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsListPageView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List News",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/events/{event_key}": {
      "get": {
        "operationId": "get_news_event_detail_api_news_events__event_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_key",
            "required": true,
            "schema": {
              "title": "Event Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsEventDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Event Detail",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/feed-layout": {
      "get": {
        "operationId": "get_news_feed_layout_api_news_feed_layout_get",
        "parameters": [
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market"
            }
          },
          {
            "in": "query",
            "name": "limit_events",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit Events",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit_topics",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit Topics",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit_stream",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit Stream",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsFeedLayoutView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Feed Layout",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/refresh": {
      "post": {
        "operationId": "refresh_news_sources_api_news_refresh_post",
        "parameters": [
          {
            "in": "query",
            "name": "async_mode",
            "required": false,
            "schema": {
              "default": false,
              "title": "Async Mode",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh News Sources",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/runtime": {
      "get": {
        "operationId": "get_news_runtime_api_news_runtime_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsRuntimeView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Runtime",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/{news_id}": {
      "get": {
        "operationId": "get_news_detail_api_news__news_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "news_id",
            "required": true,
            "schema": {
              "title": "News Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Detail",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/{news_id}/analysis": {
      "get": {
        "operationId": "get_news_analysis_api_news__news_id__analysis_get",
        "parameters": [
          {
            "in": "path",
            "name": "news_id",
            "required": true,
            "schema": {
              "title": "News Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/NewsAnalysisView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get News Analysis Api News  News Id  Analysis Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get News Analysis",
        "tags": [
          "news"
        ]
      }
    },
    "/api/news/{news_id}/analyze": {
      "post": {
        "operationId": "analyze_news_api_news__news_id__analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "news_id",
            "required": true,
            "schema": {
              "title": "News Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-App-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-App-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsAnalysisView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze News",
        "tags": [
          "news"
        ]
      }
    },
...
  }
}